refactor : step3 페이지 client - server component 분리
